refactor: extract normalizeBerechnungsart to shared utility
- Moved normalizeBerechnungsart function to a shared utility file
- Updated imports in betriebskosten-edit-modal.tsx and sortable-cost-item.tsx
- Removed duplicate function definition from sortable-cost-item.tsx
- Improved code organization and maintainability
