+++ components/betriebskosten-edit-modal.tsx
@@ -23,22 +23,8 @@ import {
 } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
 
-const normalizeBerechnungsart = (rawValue: string): BerechnungsartValue => {
-  const berechnungsartMap: Record<string, BerechnungsartValue> = {
-    'pro person': 'pro Mieter',
-    'pro mieter': 'pro Mieter',
-    'pro flaeche': 'pro Flaeche',
-    'pro fläche': 'pro Flaeche',
-    'pro qm': 'pro Flaeche',
-    'qm': 'pro Flaeche',
-    'pro wohnung': 'pro Wohnung',
-    'nach rechnung': 'nach Rechnung',
-  };
-  
-  const lower = rawValue.toLowerCase();
-  const normalized = berechnungsartMap[lower] || rawValue;
-  return (BERECHNUNGSART_OPTIONS.find(opt => opt.value === normalized)?.value as BerechnungsartValue) || '';
-};
+import { normalizeBerechnungsart } from "../utils/betriebskosten";
+
 
 import {
   Dialog,
