+++ components/sortable-cost-item.tsx
@@ -6,6 +6,7 @@ import { createPortal } from "react-dom";
 import { useSortable } from '@dnd-kit/sortable';
 import { CSS } from '@dnd-kit/utilities';
 import { PlusCircle, Trash2, GripVertical } from "lucide-react";
+import { normalizeBerechnungsart } from "../utils/betriebskosten";
 import { Button } from "./ui/button";
 import { Input } from "./ui/input";
 import { Label } from "./ui/label";
@@ -47,23 +48,6 @@ export interface SortableCostItemProps {
   onItemLeave: () => void;
   selectContentRef: React.RefObject<HTMLDivElement | null>;
 }
-
-const normalizeBerechnungsart = (rawValue: string): BerechnungsartValue => {
-  const berechnungsartMap: Record<string, BerechnungsartValue> = {
-    'pro person': 'pro Mieter',
-    'pro mieter': 'pro Mieter',
-    'pro flaeche': 'pro Flaeche',
-    'pro fläche': 'pro Flaeche',
-    'pro qm': 'pro Flaeche',
-    'qm': 'pro Flaeche',
-    'pro wohnung': 'pro Wohnung',
-    'nach rechnung': 'nach Rechnung',
-  };
-  
-  const lower = rawValue.toLowerCase();
-  const normalized = berechnungsartMap[lower] || rawValue;
-  return (BERECHNUNGSART_OPTIONS.find(opt => opt.value === normalized)?.value as BerechnungsartValue) || '';
-};
 
 export function SortableCostItem({
   item,
